% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertempr.R
\name{celsius_to_fahr}
\alias{celsius_to_fahr}
\title{Convert Celsius to Fahrenheit}
\usage{
celsius_to_fahr(temp)
}
\arguments{
\item{temp}{numeric}
}
\value{
numeric
}
\description{
Convert a temperature from Celsius to Fahrenheit
}
\examples{
celsius_to_fahr(0)
}
