% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertempr.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Convert Celsius to Kelvin}
\usage{
celsius_to_kelvin(temp)
}
\arguments{
\item{temp}{numeric}
}
\value{
numeric
}
\description{
Convert a temperature from Celsius to Kelvin
}
\examples{
celsius_to_kelvin(0)
}
