% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertempr.R
\name{fahr_to_celsius}
\alias{fahr_to_celsius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahr_to_celsius(temp)
}
\arguments{
\item{temp}{numeric}
}
\value{
numeric
}
\description{
Convert a temperature from Fahrenheit to Celsius
}
\examples{
fahr_to_celsius(32)
}
