% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertempr.R
\name{fahr_to_kelvin}
\alias{fahr_to_kelvin}
\title{Convert Fahrenheit to Kelvin}
\usage{
fahr_to_kelvin(temp)
}
\arguments{
\item{temp}{numeric}
}
\value{
numeric
}
\description{
Convert a temperature from Fahrenheit to Kelvin
}
\examples{
fahr_to_kelvin(32)
}
