% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertempr.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Convert Kelvin to Celsius}
\usage{
kelvin_to_celsius(temp)
}
\arguments{
\item{temp}{numeric}
}
\value{
numeric
}
\description{
Convert a temperature from Kelvin to Celsius
}
\examples{
kelvin_to_celsius(273.15)
}
