% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertempr.R
\name{kelvin_to_fahr}
\alias{kelvin_to_fahr}
\title{Convert Kelvin to Fahrenheit}
\usage{
kelvin_to_fahr(temp)
}
\arguments{
\item{temp}{numeric}
}
\value{
numeric
}
\description{
Convert a temperature from Kelvin to Fahrenheit
}
\examples{
kelvin_to_fahr(273.15)
}
